fix: compare the peer UID without narrowing the current user's
os.Getuid can report a value uint32 cannot hold, and the narrowing wrapped it
onto a real peer UID. Comparing in int64 is exact from both sides and drops
gosec G115 without an annotation.

+++ internal/tenantfs/authorization.go
@@ -98,7 +98,7 @@ func boundDomain(id catalog.TenantID) (catalogproto.DomainID, error) {
 }
 
 func validCatalogIdentity(uid int, identity catalogservice.Identity) bool {
-	return identity.Caller.PID > 1 && identity.Caller.UID == uint32(uid)
+	return identity.Caller.PID > 1 && int64(identity.Caller.UID) == int64(uid)
 }
 
 var (
